Show `mediaTypes.native.image.sizes` in examples
In addition to updating the basic examples to use sizes for images and
icons, we add a section to both the API reference and the 'Show Native
Ads' page that lists both ways to define sizes for image-like assets.

Since the content is the same, it's in a shared file sourced using the
Liquid `include` keyword.