Make table responsive for example
Makes the table scrollable if it’s wider than the space.
Have to wrap the markdown in `div` tags and move the Inline Attribution List (https://kramdown.gettalong.org/syntax.html#inline-attribute-lists) to the bottom so that it doesn’t interfere with Markdown parsing of the table.
Also need to specify that the `div` contains markdown via `markdown=“1”` based on https://kramdown.gettalong.org/syntax.html#html-blocks