Add helper script to find ungrouped bidders
Whenever bidders are added to the site they need to be assigned to a `list_group` from `dev-docs/bidders.md` via the individual bidder frontmatter. If a bidder isn’t assigned to a `list_group` then they won’t appear on the Bidders page.

This script checks all bidders and lists those that aren’t assigned to a `list_group`.

Exclude the script so it’s not included in the generated site.