Tappx :: avoid host (#3520)
* Tappx :: avoid host

This changes is in response to the pull request: https://github.com/prebid/prebid-server/pull/2130

* Change to be more accurate

Change thanks to Bretg in the pull request: https://github.com/prebid/prebid-server/pull/2130
I take a look to prebid.js module of tappx and we will made the same change in a near future. However, it is better to put correctly in the docs.
Thanks.