Update axis.md
added Muki's suggested overview. Combined tables, listed the iabCat parameter as JS only. Really that parameter should be dropped because we don't want pubs to have to set bidder-specific fields like this. Your adapter should be reading the ortb2 field. But this PR is way late. Just pushing it to completion.