New Adapter: Axis (#4424)
* New Adapter: Axis

* tcfeu header

as covered in a separate PR, changing gdpr_supported to tcfeu_supported and setting to false. You need to supply a GVL ID in order to cleaim TCF-EU support

* split bids params and added gvl_id

* Update axis.md

added Muki's suggested overview. Combined tables, listed the iabCat parameter as JS only. Really that parameter should be dropped because we don't want pubs to have to set bidder-specific fields like this. Your adapter should be reading the ortb2 field. But this PR is way late. Just pushing it to completion.

* linting update

* linting, cont.

---------

Co-authored-by: Maksym Pavliv <makspavliv@MacBook-Air-Maksym.local>
Co-authored-by: bretg <bgorsline@gmail.com>
Co-authored-by: Muki Seiler <muuki88@users.noreply.github.com>